Trying to make justify my CS major at the moment
